requirements_bot.txt

python-dotenv==1.2.2
requests==2.33.1
python-telegram-bot==13.0